Change old router link directive to navigation guard directive
