+++ angular-workspace/projects/example-client-app/src/app/customapp/customapp.component.html
@@ -202,9 +202,6 @@
                     <nimble-icon-check title="String 1"></nimble-icon-check>
                 </nimble-table-column-text>
                 <nimble-table-column-anchor
-                    nimbleRouterLink
-                    [queryParams]="{ word: 'foo' }"
-                    queryParamsHandling="merge"
                     href-field-name="href"
                     label-field-name="linkLabel"
                 >
