+++ angular-workspace/projects/ni/nimble-angular/table-column/anchor/nimble-table-column-anchor.module.ts
@@ -1,14 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NimbleTableColumnAnchorDirective } from './nimble-table-column-anchor.directive';
-import { NimbleTableColumnAnchorRouterLinkWithHrefDirective } from './nimble-table-column-anchor-router-link-with-href.directive';
-import { NimbleTableColumnAnchorRouterLinkDirective } from './nimble-table-column-anchor-router-link.directive';
+import { NimbleTableColumnAnchorNavigationGuardDirective } from './nimble-table-column-anchor-navigation-guard.directive';
 
 import '@ni/nimble-components/dist/esm/table-column/anchor';
 
 @NgModule({
-    declarations: [NimbleTableColumnAnchorDirective, NimbleTableColumnAnchorRouterLinkDirective, NimbleTableColumnAnchorRouterLinkWithHrefDirective],
+    declarations: [NimbleTableColumnAnchorDirective, NimbleTableColumnAnchorNavigationGuardDirective],
     imports: [CommonModule],
-    exports: [NimbleTableColumnAnchorDirective, NimbleTableColumnAnchorRouterLinkDirective, NimbleTableColumnAnchorRouterLinkWithHrefDirective]
+    exports: [NimbleTableColumnAnchorDirective, NimbleTableColumnAnchorNavigationGuardDirective]
 })
 export class NimbleTableColumnAnchorModule { }
