+++ angular-workspace/projects/ni/nimble-angular/table-column/anchor/public-api.ts
@@ -1,4 +1,3 @@
-export * from './nimble-table-column-anchor-router-link-with-href.directive';
-export * from './nimble-table-column-anchor-router-link.directive';
+export * from './nimble-table-column-anchor-navigation-guard.directive';
 export * from './nimble-table-column-anchor.directive';
 export * from './nimble-table-column-anchor.module';
